requirements.txt
dj-database-url==0.4.2
Django==2.0.1
django-model-utils==3.1.1
django-sb-admin==0.3.6
python-decouple==3.1
pytz==2017.3
Unipath==1.1
